Hidden nav

+++ ByvUtils/Classes/Date+extension.swift
@@ -9,14 +9,14 @@
 import Foundation
 
 public extension Date {
-    public static func fromMilliseconds(_ milliseconds: Int?) -> Date? {
+    public static func fromMilliseconds(_ milliseconds: Double?) -> Date? {
         if let millis = milliseconds {
             return Date(timeIntervalSince1970: Double(millis/1000))
         }
         return nil
     }
     
-    public static func fromMilliseconds(_ milliseconds: Int) -> Date {
+    public static func fromMilliseconds(_ milliseconds: Double) -> Date {
         return Date(timeIntervalSince1970: Double(milliseconds/1000))
     }
